+++ src/main/java/pl/team/carrent/employee/LoginController.java
@@ -1,4 +1,4 @@
-package pl.team.carrent.login;
+package pl.team.carrent.employee;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
